rename: v2 폴더 구조에 따른 이동
