+++ src/main/java/com/kau/capstone/v2/vaccination/dto/OneVaccinationResV2.java
@@ -1,4 +1,4 @@
-package com.kau.capstone.v1.vaccination.dto;
+package com.kau.capstone.v2.vaccination.dto;
 
 import com.kau.capstone.entity.vaccination.Vaccination;
 import java.time.LocalDate;
